# ExecutionNode.Path optimization v2