--- a/src/GraphQL/Execution/ExecutionHelper.cs
+++ b/src/GraphQL/Execution/ExecutionHelper.cs
@@ -445,5 +445,4 @@ namespace GraphQL.Execution
             return CollectFields(context, fieldType, field.SelectionSet);
         }
     }
-
 }