A proposed update pertaining to #1147 and replacing #1505 .

Benchmarks:
```
// * Summary *

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18362
Intel Core i7-7740X CPU 4.30GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.1.101
  [Host]     : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X64 RyuJIT
  DefaultJob : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X64 RyuJIT

*** BEFORE ***

|        Method |        Mean |     Error |    StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------- |------------:|----------:|----------:|---------:|--------:|------:|----------:|
| Introspection | 1,483.73 us | 12.154 us | 10.774 us | 191.4063 | 72.2656 |     - | 943.57 KB |
|          Hero |    31.96 us |  0.247 us |  0.219 us |   6.1035 |       - |     - |  25.21 KB |

*** AFTER ***

|        Method |        Mean |    Error |   StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------- |------------:|---------:|---------:|---------:|--------:|------:|----------:|
| Introspection | 1,354.75 us | 6.971 us | 6.180 us | 189.4531 | 62.5000 |     - | 835.55 KB |
|          Hero |    31.53 us | 0.348 us | 0.326 us |   6.1035 |       - |     - |   25.1 KB |
```

In order to validate new code design, I first added the new code without removing the old code, and temporarily used this code within the constructor of ExecuteNode to validate the new design:
``` C#
if (!Enumerable.SequenceEqual(this.Path, this.Path2))
{
    throw new Exception("Invalid path");
}
```
Everything worked, so the old code was removed.  All tests pass, of course.