@Shane32 2020-02-13
Also note that it could be programmed to use a yield iterator, but since the results need to be reversed, the list will need to be enumerated before it can be returned anyway.  See sample below:

``` C#
        public IEnumerable<string> Path => ReversePath().Reverse();

        private IEnumerable<string> ReversePath()
        {
            var node = this;
            while (!(node is RootExecutionNode))
            {
                if (node.PathIndex.HasValue)
                    yield return GetStringIndex(node.PathIndex.Value);
                else
                    yield return node.Field.Name;
                node = node.Parent;
            }
        }
```

You could walk the nodes' parents first and reverse them before yielding the names, but then you're creating the list ahead of time also.  So I just compiled the list, reversed and returned it.