@sungam3r 2020-02-13
@Shane32 I really like this PR. In fact when I told you in https://github.com/graphql-dotnet/graphql-dotnet/pull/1505#issuecomment-575250542 that I want to change this place in a complete different way I wanted to do just that. But as often happens, I put off this work because I had to do something else.