@sungam3r 2020-02-13
Also I think that we need the better name for `int? pathIndex` to point that it is index of **current** node inside parent node **if** parent node is array, otherwise `null`.