@Shane32 2020-02-13
I'm glad you like it.  Not trying to step on your toes, but just trying to help get the 3.0 milestones done.  Honestly I'd like to see a non-preview 3.0 release, even if that means the very next version is 4.0 preview 1.  Also hoping my dataloader PR gets pulled in, as I really think it solves a number of bugs and works much better for the community.  (And I use my custom dataloader code all the time, so of course I think it's great!)