@Shane32 2020-02-13
Give me a suggestion and I'll rename it, no big deal.  It was just the first thing that made sense to me.  I was going to have it protected, but then it couldn't be accessed by the subscription code that duplicated the node.  Besides, probably not a bad idea to have it accessible.