@sungam3r 2020-02-13
1. I do not have permissions to publish this package to nuget.
2. The problem with the dataloder is that, basically, I can’t afford to merge code that I don’t understand completely. I have been conducting a review for quite long time because of a detailed analysis of the code. Dataloder and subscriptions are two things that I don’t have time to deal with, so many PRs are still open. But I do not give up hope to change this trend.