@Shane32 2020-02-13
No I understand both points completely.  I think joemcbride wants to look at it himself, based on his comment, actually.  He's probably a busy person.  And my suggested changes to the dataloader are not insignificant.