@sungam3r 2020-02-13
I propose `IndexInParentArrayExecutionNode`. It seems rather long name though.