@Shane32 2020-02-13
Yeah ..... how about simplifying it just a bit -- `IndexInParent` or `IndexInParentArray` or `IndexInParentNode` ?  Maybe reverse the name and remove 'in' e.g. `ParentNodeIndex` ?

We can do the long name, but it's ... quite long ... even for descriptive variable names.  What do you think?