@Shane32 2020-02-13
`ParentNodeIndex` doesn't seem clear enough to me.  I'd favor `IndexInParentArray` or possibly `IndexInParentNode`