@Shane32 2020-02-13
It could, but it's a very quick calculation, and considering how infrequent Path (should) be used, I'd save the memory in ExecutionNode.  I can add it if you think that would be better.