@sungam3r 2020-02-13
Results of our work. No more `String[]` in top of memory traffic:
![image](https://user-images.githubusercontent.com/21261007/74481346-9b574b00-4ec3-11ea-9fab-d34ce61459a8.png)